Refactor: Split server performance tests into logical files
Split the 816-line server_perf_test.go into 4 focused files for better
organization and maintainability:

**server_perf_test.go** (175 lines):
- Shared helper functions (setupBenchmarkServers, setupBenchmarkServersWithSize)
- BenchmarkExport (used across all views)

**server_perf_table_test.go** (169 lines):
- BenchmarkTable
- BenchmarkLargeResultSets
- BenchmarkFilterHeavyTableView
- BenchmarkConcurrentTableView

**server_perf_topology_test.go** (202 lines):
- BenchmarkTopologyLoki
- BenchmarkTopologyAuto
- BenchmarkFilterHeavyTopology
- BenchmarkConcurrentTopology
- BenchmarkTopologyAggregations
- runTopologyQuery helper

**server_perf_overview_test.go** (288 lines):
- BenchmarkOverviewLoki
- BenchmarkOverviewAuto
- BenchmarkFilterHeavyOverview
- BenchmarkConcurrentOverview
- BenchmarkOverviewAggregations
- Overview query builder helpers

All benchmarks verified and working correctly.
